backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-google-genai
python-dotenv
numpy
